requirements.txt
langchain
langchain-text-splitters
langchain-huggingface
langchain-community
langchain-core
sentence-transformers
faiss-cpu
groq
python-dotenv
torch